Move helpers into common folder
