+++ lib/plugin-downgrade-helpers/downgrade-burst-shutter-effect.js
