+++ lib/plugin-downgrade-helpers/replace-null-switch-channels.js
